+++ packages/ui/src/styles/FlexBox.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-type TFlexBox = {
+export type TFlexBox = {
   direction?: 'row' | 'row-reverse' | 'column' | 'column-reverse'
   justify?:
     | 'flex-start'
@@ -13,10 +13,28 @@ type TFlexBox = {
   gap?: string
 }
 
-export const FlexBox = styled.div<TFlexBox>`
+// const Comp = styled('div').withConfig({
+//   shouldForwardProp: (prop) =>
+//       !['hidden'].includes(prop),
+// }).attrs({ className: 'foo' })`
+//   color: red;
+//   &.foo {
+//     text-decoration: underline;
+//   }
+// `;
+export const FlexBox = styled.div.withConfig<TFlexBox>(
+{ shouldForwardProp: (prop) => !['direction', 'justify', 'align', 'gap']
+.includes(prop) })<TFlexBox>`
   display: flex;
   flex-direction: ${({ direction }) => direction || 'column'};
   justify-content: ${({ justify }) => justify || 'center'};
   align-items: ${({ align }) => align || 'center'};
   gap: ${({ gap }) => gap || '0'};
-`
+
+// export const FlexBox = styled.div<TFlexBox>`
+//   display: flex;
+//   flex-direction: ${({ direction }) => direction || 'column'};
+//   justify-content: ${({ justify }) => justify || 'center'};
+//   align-items: ${({ align }) => align || 'center'};
+//   gap: ${({ gap }) => gap || '0'};
+// `